pyproject.toml
[project]
name = "make-money-easy"
version = "0.1.0"
description = "Taiwan stock paper trading app built with FastAPI, SQLite, and twstock."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "apscheduler>=3.10.4,<4.0.0",
    "email-validator>=2.2.0,<3.0.0",
    "fastapi>=0.115.0,<1.0.0",
    "pydantic-settings>=2.6.1,<3.0.0",
    "sqlalchemy>=2.0.36,<3.0.0",
    "twstock>=1.4.0,<2.0.0",
    "uvicorn[standard]>=0.32.0,<1.0.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28.1,<1.0.0",
    "pytest>=8.3.4,<9.0.0",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
